requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Evaluation framework
lm_eval>=0.4.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Logging and tracking
wandb>=0.16.0
tensorboard>=2.14.0
tqdm>=4.66.0

# API and web
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data processing
datasets>=2.15.0
pyarrow>=14.0.0
tokenizers>=0.15.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
typer>=0.9.0
jinja2>=3.1.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Optional: llama.cpp Python bindings
# llama-cpp-python>=0.2.0